requirements.txt
django==4.1.6
djangorestframework==3.14.0
django-rest-passwordreset==1.3.0
psycopg2-binary==2.9.5
python-dotenv==0.21.1
pytest==7.2.1
pytest-django==4.5.2
model-bakery==1.10.1
celery==5.2.7
requests==2.28.2
redis==4.5.1
pyyaml==6.0
